feat:迁移各适配器配置项至对应适配器下
